requirements.txt
torch==1.8.2+cu111 -f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torchvision==0.9.2+cu111 -f https://download.pytorch.org/whl/lts/1.8/torch_lts.html
torchaudio===0.8.2
torchsummary
pillow
numpy
opencv-python
scikit-learn